Use a transaction in createProject in production
Do not use a transaction when running in the dev server as this causes a
multi-entity group error. This happens because the dev server simulates
the Google Cloud Store using the datastore. If we use a transaction in
create project it winds up covering multiple entity groups. One for our
modifications to the “FileData” kind and a second for the simulated GCS
bucket kind.

Change-Id: I4a31eb16c7866e4ed721c02dbc77addea8df3dbf