Prevent saving projects settings during load
When a project is loaded, the MockForm for Screen1 sets the AppName
and other project settings. This has the side effect of immediately
scheduling an autosave.

This commit changes the behavior so that if Screen1 is still in a
loading state (i.e., no editor is active or the active editor is the
YaFormEditor but the MockForm doesn't exist), then the autosave will
not be triggered. Only after MockForm exists (and thus its properties
have all been set) will an autosave be triggered in response to
changing a project setting.

Change-Id: I51a9b2607bcf8150f42fe553da74d4e7b5bf47da